// src/Utils/types.ts
export type IDType = string | number;

export interface ImageType {
  src?: any;
  alt?: string;
  width?: number;
  height?: number;
  loading?: "lazy" | "eager";
}

export interface HeadingType {
  id?: IDType;
  content?: string;
}

export interface TextType {
  id?: IDType;
  content?: string;
}

export interface ButtonType {
  id?: IDType;
  content?: string;
  path?: string;
  icon?: string;
}

export interface SectionTitleType {
  title?: string;
  subtitle?: string;
}

export interface AnchorType {
  id?: IDType;
  path?: string;
  content?: string;
}

export interface VideoType {
  videoId?: string;
}

export interface MottoType {
  text?: string;
  path?: string;
  pathText?: string;
}

export interface ItemType {
  id?: IDType;
  title?: string;
  description?: string;
  icon?: string;
  path?: string;
  pathText?: string;
  suffix?: string;
  counter?: number;
  name?: string;
  designation?: string;
  rating?: number;
  role?: "admin" | "employee";
  headings?: HeadingType[];
  texts?: TextType[];
  images?: ImageType[];
  motto?: MottoType;
}
// ...
export type TSection = {
  space?:
    | "top-bottom"
    | "top-bottom-2"
    | "top-bottom-3"
    | "top"
    | "top-2"
    | "bottom"
    | "bottom-2"
    | "bottom-3"
    | "none";
  bg?: string;
  titleSize?: "default" | "large";
  className?: string;
};